# src/main/java/com/pragma/pagos/interfaces/PagoController.java
package com.pragma.pagos.interfaces;

import jakarta.validation.Valid;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

import java.util.Map;

@RestController
@RequestMapping("/api/pagos")
public class PagoController {

    @PostMapping
    public ResponseEntity<Map<String, Object>> crearPago(
            @Valid @RequestBody Map<String, Object> request,
            @RequestHeader(value = "X-Correlation-ID", required = false) String correlationId) {
        return ResponseEntity.status(201).body(Map.of(
                "exito", true,
                "correlationId", correlationId != null ? correlationId : java.util.UUID.randomUUID().toString(),
                "solicitud", request
        ));
    }

    @GetMapping("/health")
    public ResponseEntity<Map<String, String>> health() {
        return ResponseEntity.ok(Map.of("status", "UP"));
    }
}